% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cogdata.R
\name{make.cogdata}
\alias{make.cogdata}
\title{Parse orthogroups tabular output from OrthoFinder into a `cogdata` data frame for geneplast}
\usage{
make.cogdata(file)
}
\arguments{
\item{file}{OrthoFinder orthogroups tabular file}
}
\value{
cogdata data frame
}
\description{
Parse orthogroups tabular output from OrthoFinder into a `cogdata` data frame for geneplast
}
\author{
Leonardo RS Campos
}
