% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_phylo_tree.R
\name{make.phyloTree}
\alias{make.phyloTree}
\title{Build a `phyloTree` object for geneplast}
\usage{
make.phyloTree(sspids = NULL, newick = NULL, verbose = TRUE)
}
\arguments{
\item{sspids}{a vector or data frame containing NCBI Taxon IDs from the species of interest.}

\item{newick}{a rooted phylogenetic tree in Newick format.}

\item{verbose}{a logical value specifying whether or not to display detailed messages.}
}
\value{
An object of class "phylo".
}
\description{
Build a `phyloTree` object for geneplast
}
\author{
Danilo O Imparato
}
